test script modified (compare graphs too)

+++ Transformations/TaskGraph/tests/test.sh
@@ -15,20 +15,26 @@ expectedPath="$curPath/expected"
 resultPath="$curPath/results"
 testsPath="$curPath/autotests"
 
+rawGraphName="taskGraphRaw.dot"
+optimizedGraphName="taskGraphOpt.dot"
+
 mkdir -p $resultPath
 # Iterate over all .cpp files in the current directory
 for file in $testsPath/*.cpp; do
+    fileName=$(basename "$file" /)
+    folderName=$(basename "$fileName" .cpp)
+    folderResultPath="$resultPath/$folderName"
+
     # Check if the file exists and is a regular file
     if [ -f "$file" ]; then
         ((countTotal++))
         differenceInText=false
         differenceInAST=false
-        fileName=$(basename "$file" /)
-        folderName=$(basename "$fileName" .cpp)
-        mkdir $resultPath/$folderName
-        folderResultPath="$resultPath/$folderName"
-        $taskGraph $file $arguments > "$folderResultPath/$fileName" 2> /dev/null
-                         
+        differenceInRawGraph=false
+        differenceInOptimizedGraph=false
+
+        mkdir -p $resultPath/$folderName
+        $taskGraph $file $arguments > "$folderResultPath/$fileName" 2> /dev/null             
         if [ $? -ne 0 ]; then
             echo -e "${RED}${BOLD}Failed to process file : $file"
             differenceInAST=true
@@ -42,24 +48,36 @@ for file in $testsPath/*.cpp; do
                     diff "$file2" "$expectedResult" > /dev/null
                     if [ $? -ne 0 ]; then
                         differenceInText=true
-                        echo -e "${RED}Different Text : $fileName${NC}"
                     fi
                     clang-check -ast-print "$expectedResult" > "$curPath/ast_result" 2> /dev/null
                     clang-check -ast-print $file2 > "$curPath/ast_expected" 2> /dev/null
                     diff "$curPath/ast_expected" "$curPath/ast_result" > "$folderResultPath"/astdiff
                     if [ $? -ne 0 ]; then
                         differenceInAST=true
-                        echo -e "${RED}Different AST : $fileName${NC}"
                     fi
+                    
                 fi
             done
             rm "$curPath/ast_expected"
             rm "$curPath/ast_result"
-        fi 
+        fi
+        if [ -f "$rawGraphName" ]; then
+            mv $rawGraphName "$folderResultPath"
+            diff "$folderResultPath/$rawGraphName" "$expectedPath/$rawGraphName" > /dev/null
+                if [ $? -ne 0 ]; then
+                    differenceInRawGraph=true
+                fi
+        fi
+        if [ -f "$optimizedGraphName" ]; then
+            mv $optimizedGraphName "$folderResultPath"
+            diff "$folderResultPath/$optimizedGraphName" "$expectedPath/$optimizedGraphName" > /dev/null
+                if [ $? -ne 0 ]; then
+                    differenceInOptimizedGraph=true
+                fi
+        fi
         if $differenceInAST; then
             echo -e "${RED}${BOLD}Test failed, AST : $folderName${NC}"
         else 
-            ((countPassed++))
             echo -e "${GREEN}${BOLD}Test succeeded : $folderName${NC}"
         fi
         if $differenceInText; then
@@ -67,7 +85,18 @@ for file in $testsPath/*.cpp; do
         else 
             echo -e "${GREEN}Test succeeded, TEXT : $folderName${NC}"
         fi
-        if [ $differenceInAST == false ]; then #&& [ $differenceInText == false ]
+        if $differenceInRawGraph; then
+            echo -e "${RED}Test failed, RAW GRAPH : $folderName${NC}"
+        else 
+            echo -e "${GREEN}Test succeeded, RAW GRAPH : $folderName${NC}"
+        fi
+        if $differenceInOptimizedGraph; then
+            echo -e "${RED}Test failed, OPTIMIZED GRAPH : $folderName${NC}"
+        else 
+            echo -e "${GREEN}Test succeeded, OPTIMIZED GRAPH : $folderName${NC}"
+        fi
+        if [ $differenceInAST == false ] && [ $differenceInText == false ] && [ $differenceInRawGraph == false ] && [ $differenceInOptimizedGraph == false ]; then
+            ((countPassed++)) 
             rm -rf "$folderResultPath"
         fi
     fi
